requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.0.0
gevent>=23.0.0
requests>=2.31.0
python-dotenv>=1.0.0
redis>=5.0.0
apscheduler>=3.10.0
pytest>=7.0.0
pytest-cov>=4.0.0
